added automapper. Changed price to decimal. Fixed some issues on the basket controller. Changed to nunit. Added response entities on the web api

+++ Checkout.Entities/Validator/ItemValidator.cs
@@ -9,7 +9,7 @@ namespace Checkout.Entities.Validator
         {
             RuleFor(reg => reg.Name).NotEmpty().WithMessage("Name cannot be empty");
             RuleFor(reg => reg.Description).NotEmpty().WithMessage("Desciption cannot be empty");
-            RuleFor(reg => reg.Price).NotEmpty().ExclusiveBetween(0, float.MaxValue).WithMessage("Price must be greater than 0");
+            RuleFor(reg => reg.Price).NotEmpty().ExclusiveBetween(0, decimal.MaxValue).WithMessage("Price must be greater than 0");
         }
     }
 }
